Move jsrepl to inside of /public/assets to skip assets pipeline.
Jsrepl is too big/complicated/hardcore for assets pipeline to handle.
On Heroku, deployment will ALWAYS fail if jsrepl is handled by assets pipeline.

I proceed to try out S3 since it's fast and cheap, but jsrepl is using
[web worker](https://developer.mozilla.org/en/docs/Web/Guide/Performance/Using_web_workers)
and that requires the resources to be obey the same-origin policy. But clearly,
resources coming from S3 will definitely not obey this policy unless we
implement some hackery.

So, I think this method should work. :D