Implement a hook to render React components
Elements with data-react-component specified will be rendered with
the corrent React component along with specified props.

The names of the props will be automatically tranformed to conform
to React's convention, that is, camel case instead of snake case.

Note that this implementation is not perfect, but should be enough
for this project. For example, data-react-data-my-attr should be
passed as data-my-attr but is passed as dataMyAttr.